+++ revdep/check.R
@@ -4,37 +4,39 @@ devtools::install_github("r-lib/revdepcheck")
 library(revdepcheck)
 
 revdep_reset()
-revdep_check(num_workers = 6)
+revdep_check(num_workers = 4)
 
 revdep_summary()
 revdep_todo()
 
 revdep_add_broken(install_failures = F, timeout_failures = T)
-revdep_check(num_workers = 6)
+revdep_check(num_workers = 4)
 
 ## revdep_rm(packages = "xROI")
 ## revdep_add(packages = "GenEst")
 revdep_check()
 
+(smr <- revdep_summary())
+revdep_details(revdep = "GenEst")
 
-revdep_summary()
-revdep_report()
-revdep_report_problems()
+## in another process
 revdep_report_summary()
-
+revdep_report_problems()
 revdep_summary()
 revdep_maintainers()
 
-
+## finally
+revdep_report()
 
 
 db_delete <- function(pkgdir, package) {
-  db <- db(pkgdir)
-  dbExecute(db,
-    sqlInterpolate(db,
+  db <- revdepcheck:::db(pkgdir)
+  DBI:::dbExecute(db,
+    DBI::sqlInterpolate(db,
       "DELETE FROM revdeps WHERE package = ?package",
       package = package
     )
   )
 }
-## revdepcheck:::db_delete(".", "GenEst")
+
+db_delete(".", "GenEst")