+++ tests/testthat/test-update.R
@@ -6,7 +6,9 @@ test_that("update.Date returns a date object", {
   expect_s3_class(update(date, wdays = 1), "Date")
   expect_s3_class(update(date, months = 1), "Date")
   expect_s3_class(update(date, years = 2001), "Date")
-  expect_s3_class(update(date, tzs = "UTC"), "Date")
+  ## expect_s3_class(update(date, tzs = "UTC"), "POSIXct")
+  tz(date) <- "UTC"
+  expect_s3_class(date, "POSIXct")
 })
 
 test_that("update.Date returns a posix object if time is manipulated", {
@@ -375,6 +377,36 @@ test_that("update returns NA for date-times in the spring dst gap", {
   expect_true(is.na(update(poslt, tzs = "America/New_York")))
 })
 
+test_that("update handles repeated DST transitions", {
+  repref <- ymd_hms("2022-10-30 01:00:00", tz = "Europe/Amsterdam")
+  rep <- ymd_hms("2022-10-29 00:01:02", tz = "Europe/Amsterdam")
+
+  ## expect_equal(update(rep, days = 30, hours = 2, minutes = 3, roll_dst = "pre"),
+  ##   repref + 3600 + 3*60 + 2)
+  ## expect_equal(update(rep, days = 30, hours = 2, minutes = 3, roll_dst = "boundary"),
+  ##   repref + 2*3600)
+  ## expect_equal(update(rep, days = 30, hours = 2, minutes = 3, roll_dst = "post"),
+  ##   repref + 2*3600 + 3*60 + 2)
+  ## expect_equal(update(rep, days = 30, hours = 2, seconds = 3.35, roll_dst = "pre"),
+  ##   repref + 3600 + 60 + 3.35)
+  ## expect_equal(update(rep, days = 30, hours = 2, seconds = 3.35, roll_dst = "boundary"),
+  ##   repref + 2*3600)
+  ## expect_equal(update(rep, days = 30, hours = 2, seconds = 3.35, roll_dst = "post"),
+  ##   repref + 2*3600 + 60 + 3.35)
+
+
+  ## our default roll_dst is post
+  repref <- ymd_hms("2022-10-30 01:00:00", tz = "Europe/Amsterdam")
+  rep <- ymd_hms("2022-10-29 00:10:02.35", tz = "Europe/Amsterdam")
+  expect_equal(update(rep, days = 30, hours = 2, minutes = 3),
+    repref + 2*3600 + 3*60 + 2.35)
+  expect_equal(update(rep, days = 30, hours = 2, seconds = 5.6),
+    repref + 2*3600 + 10*60 + 5.6)
+  expect_equal(update(rep, days = 30, hours = 2, seconds = 65.6),
+    repref + 2*3600 + 11*60 + 5.6)
+
+})
+
 test_that("update handles vectors of dates", {
   poslt <- as.POSIXlt(c("2010-02-14 01:59:59", "2010-02-15 01:59:59", "2010-02-16 01:59:59"),
     tz = "UTC", format = "%Y-%m-%d %H:%M:%S"
@@ -455,6 +487,17 @@ test_that("Updateing on seconds doesn't affect hours", {
   expect_equal(hour(tt), hour(tt2))
 })
 
+test_that("Subtracting months to March 1 produces correct results", {
+  ## #1037
+  time <- ymd("2022-04-01", tz = "America/New_York")
+  expect_equal(time - months(1), ymd("2022-03-01", tz = "America/New_York"))
+  time <- ymd("2021-04-01", tz = "America/New_York")
+  expect_equal(time - months(1), ymd("2021-03-01", tz = "America/New_York"))
+  time <- ymd("2022-05-01", tz = "America/New_York")
+  expect_equal(time - months(2), ymd("2022-03-01", tz = "America/New_York"))
+  time <- ymd_hms("2022-04-02 04:01:01", tz = "America/New_York")
+  expect_equal(time - period("1m 1d 4H 1M 1S"), ymd("2022-03-01", tz = "America/New_York"))
+})
 
 ## ## bug #319
 ## x <- structure(list(sec = 0, min = 0, hour = 0, mday = -212, mon = 7L,